+++ src/common/utxobased/info/bitcoingold.ts
@@ -73,7 +73,7 @@ export const coinInfo: CoinInfo = {
   coinType: 156,
 
   prefixes: {
-    messagePrefix: ['\x18Bitcoin Gold Signed Message:\n'],
+    messagePrefix: ['\x1dBitcoin Gold Signed Message:\n'],
     wif: [0x80],
     legacyXPriv: [0x0488ade4],
     legacyXPub: [0x0488b21e],